+++ build_config/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.4.3-dev
 
+- Added the `public_assets` option, which describes the assets readable when
+  the configured package is not the root of the build.
+
 ## 0.4.2
 
 - Add support for an `auto_apply_builders` option to the `target` config.